cleanup package
